迁移主项目的RodNet的torch版代码到测试项目中去，去掉主项目的torch功能支持（仅涉及钓鱼）
