+++ BetterGenshinImpact/BetterGenshinImpact.csproj
@@ -90,7 +90,6 @@
 <!--		<PackageReference Include="supabase-csharp" Version="0.16.2" />-->
 		<PackageReference Include="System.Drawing.Common" Version="10.0.7" />
 		<PackageReference Include="System.IO.Hashing" Version="9.0.4" />
-		<PackageReference Include="TorchSharp" Version="0.105.0" />
 		<PackageReference Include="Vanara.PInvoke.CoreAudio" Version="4.1.3" />
 		<PackageReference Include="Vanara.PInvoke.NtDll" Version="4.1.3" />
 		<PackageReference Include="Vanara.PInvoke.SHCore" Version="4.1.3" />
