+++ BetterGenshinImpact/GameTask/AutoFishing/AutoFishingConfig.cs
@@ -1,4 +1,4 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.ComponentModel;
 using System;
 
 namespace BetterGenshinImpact.GameTask.AutoFishing;
@@ -46,10 +46,4 @@ public partial class AutoFishingConfig : ObservableObject
     /// </summary>
     [ObservableProperty]
     private FishingTimePolicy _fishingTimePolicy = FishingTimePolicy.All;
-
-    /// <summary>
-    /// torch库文件地址
-    /// </summary>
-    [ObservableProperty]
-    private string _torchDllFullPath = @"C:\torch\lib\torch_cpu.dll";
 }
