+++ BetterGenshinImpact/GameTask/AutoFishing/AutoFishingTaskParam.cs
@@ -2,24 +2,18 @@ using BetterGenshinImpact.GameTask.Model;
 using BetterGenshinImpact.Helpers;
 using Microsoft.ClearScript;
 using Microsoft.Extensions.Localization;
-using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Runtime.InteropServices;
-using System.Text;
-using TorchSharp;
 
 namespace BetterGenshinImpact.GameTask.AutoFishing
 {
     public class AutoFishingTaskParam : BaseTaskParam<AutoFishingTask>
     {
-        public AutoFishingTaskParam(int wholeProcessTimeoutSeconds, int throwRodTimeOutTimeoutSeconds, FishingTimePolicy fishingTimePolicy, bool saveScreenshotOnKeyTick, bool useTorch, CultureInfo? cultureInfo, IStringLocalizer<AutoFishingTask>? stringLocalizer) : base(cultureInfo, stringLocalizer)
+        public AutoFishingTaskParam(int wholeProcessTimeoutSeconds, int throwRodTimeOutTimeoutSeconds, FishingTimePolicy fishingTimePolicy, bool saveScreenshotOnKeyTick, CultureInfo? cultureInfo, IStringLocalizer<AutoFishingTask>? stringLocalizer) : base(cultureInfo, stringLocalizer)
         {
             WholeProcessTimeoutSeconds = wholeProcessTimeoutSeconds;
             ThrowRodTimeOutTimeoutSeconds = throwRodTimeOutTimeoutSeconds;
             FishingTimePolicy = fishingTimePolicy;
             SaveScreenshotOnKeyTick = saveScreenshotOnKeyTick;
-            UseTorch = useTorch;
         }
 
 
@@ -27,7 +21,6 @@ namespace BetterGenshinImpact.GameTask.AutoFishing
         public int ThrowRodTimeOutTimeoutSeconds { get; set; }
         public FishingTimePolicy FishingTimePolicy { get; set; }
         public bool SaveScreenshotOnKeyTick { get; set; }
-        public bool UseTorch { get; set; }
 
         /// <summary>
         /// 从JS请求参数构建任务参数
@@ -49,22 +42,7 @@ namespace BetterGenshinImpact.GameTask.AutoFishing
             var fishingTimePolicy = (FishingTimePolicy)ScriptObjectConverter.GetValue(jsObject, "fishingTimePolicy", (int)autoFishingConfig.FishingTimePolicy);
             var saveScreenshotOnKeyTick = ScriptObjectConverter.GetValue(jsObject, "saveScreenshotOnKeyTick", false);
 
-            bool useTorch;
-            try
-            {
-                NativeLibrary.Load(autoFishingConfig.TorchDllFullPath);
-                if (torch.TryInitializeDeviceType(DeviceType.CUDA))
-                {
-                    torch.set_default_device(new torch.Device(DeviceType.CUDA));
-                }
-                useTorch = true;
-            }
-            catch (Exception e) when (e is DllNotFoundException || e is NotSupportedException)
-            {
-                useTorch = false;
-            }
-
-            return new AutoFishingTaskParam(wholeProcessTimeoutSeconds, throwRodTimeOutTimeoutSeconds, fishingTimePolicy, saveScreenshotOnKeyTick, useTorch, null, null);
+            return new AutoFishingTaskParam(wholeProcessTimeoutSeconds, throwRodTimeOutTimeoutSeconds, fishingTimePolicy, saveScreenshotOnKeyTick, null, null);
         }
 
         /// <summary>
@@ -76,21 +54,7 @@ namespace BetterGenshinImpact.GameTask.AutoFishing
         public static AutoFishingTaskParam BuildFromConfig(AutoFishingConfig config, bool saveScreenshotOnKeyTick = false)
         {
             CultureInfo cultureInfo = new CultureInfo(TaskContext.Instance().Config.OtherConfig.GameCultureInfoName);
-            bool useTorch;
-            try
-            {
-                NativeLibrary.Load(config.TorchDllFullPath);
-                if (torch.TryInitializeDeviceType(DeviceType.CUDA))
-                {
-                    torch.set_default_device(new torch.Device(DeviceType.CUDA));
-                }
-                useTorch = true;
-            }
-            catch (Exception e) when (e is DllNotFoundException || e is NotSupportedException)
-            {
-                useTorch = false;
-            }
-            return new AutoFishingTaskParam(config.WholeProcessTimeoutSeconds, config.AutoThrowRodTimeOut, config.FishingTimePolicy, saveScreenshotOnKeyTick, useTorch, cultureInfo, null);
+            return new AutoFishingTaskParam(config.WholeProcessTimeoutSeconds, config.AutoThrowRodTimeOut, config.FishingTimePolicy, saveScreenshotOnKeyTick, cultureInfo, null);
         }
     }
 }
