+++ BetterGenshinImpact/GameTask/AutoFishing/RodInput.cs
@@ -1,4 +1,3 @@
-﻿using static TorchSharp.torch;
 
 namespace BetterGenshinImpact.GameTask.AutoFishing;
