+++ BetterGenshinImpact/GameTask/AutoFishing/RodNet.cs
@@ -1,10 +1,5 @@
-﻿using System;
+using System;
 using System.Linq;
-using static TorchSharp.torch.nn;
-using static TorchSharp.torch;
-using TorchSharp.Modules;
-using TorchSharp;
-using System.Collections.Generic;
 
 namespace BetterGenshinImpact.GameTask.AutoFishing;
 
@@ -44,18 +39,18 @@ namespace BetterGenshinImpact.GameTask.AutoFishing;
 /// 
 /// 然后就可视化一下onehot在不同距离的结果 加一个offset使得模型输出的结果在保证可以predict距离正好的结果的同时距离范围尽可能小
 /// </summary>
-public class RodNet : Module<Tensor, Tensor, Tensor, Tensor, Tensor>
+public class RodNet
 {
     const double alpha = 1734.34 / 2.5;
     // fitted parameters
-    static readonly double[] dz = {1.0307939, 1.5887239,  1.4377865, 0.8548809,
+    public static readonly double[] dz = {1.0307939, 1.5887239,  1.4377865, 0.8548809,
                                    1.8640924, -0.1687729, 1.8621461, 0.7167622,
                                    1.7071064, 1.8727832,  0.5531539};
-    static readonly double[] h_coeff = {0.5840698,  0.8029298,  0.6090596,
+    public static readonly double[] h_coeff = {0.5840698,  0.8029298,  0.6090596,
                                         -0.1390072, 0.7214464,  -0.6076725,
                                         0.3286690,  -0.2991239, 0.6072225,
                                         0.7662407,  -0.3689651};
-    static readonly double[,] weight = {{0.7779633, -1.7124480, 2.7366412},
+    public static readonly double[,] weight = {{0.7779633, -1.7124480, 2.7366412},
                                           {-0.0381155, -1.6536976, 3.5904298},
                                           {0.1947731, -0.0445049, 0.8416666},
                                           {-0.0331017, -1.3641578, 1.2834741},
@@ -66,7 +61,7 @@ public class RodNet : Module<Tensor, Tensor, Tensor, Tensor, Tensor>
                                           {-0.3597502, -1.7422760, 1.4354013},
                                           {-0.0578425, -2.0274212, 1.7173727},
                                           {-0.1225260, -1.0630554, 1.2958838}};
-    static readonly double[,] bias = {{3.1733532, 9.3601589, -11.0612173},
+    public static readonly double[,] bias = {{3.1733532, 9.3601589, -11.0612173},
                                         {6.4961057, 11.2683334, -13.7752209},
                                         {2.3662698, 2.4709859, -2.5402584},
                                         {2.4701204, 8.5112562, -7.6070199},
@@ -79,26 +74,7 @@ public class RodNet : Module<Tensor, Tensor, Tensor, Tensor, Tensor>
                                         {5.2460732, 7.7711511, -7.5998945}};
 
 
-    static readonly double[] offset = { 0.8, 0.4, 0.35, 0.35, 0.6, 0.3, 0.3, 0.8, 0.8, 0.8, 0.8 };
-
-    private Parameter thetaParameter;
-    private Parameter bParameter;
-    private Parameter dzParameter;
-    private Parameter hCoeffParameter;
-
-    public RodNet() : base("RodNet")
-    {
-        long num_embeddings = RodNet.weight.GetLength(0);
-        long embedding_dim = 3;
-
-        this.thetaParameter = new Parameter(torch.randn(num_embeddings, embedding_dim, dtype: ScalarType.Float64));
-        this.bParameter = new Parameter(torch.randn(num_embeddings, embedding_dim, dtype: ScalarType.Float64));
-
-        this.dzParameter = new Parameter(torch.zeros(num_embeddings, 1, dtype: ScalarType.Float64));
-        this.hCoeffParameter = new Parameter(torch.zeros(num_embeddings, 1, dtype: ScalarType.Float64));
-
-        RegisterComponents();
-    }
+    public static readonly double[] offset = { 0.8, 0.4, 0.35, 0.35, 0.6, 0.3, 0.3, 0.8, 0.8, 0.8, 0.8 };
 
     static void Softmax(double[] dst, double[] x, int n)
     {
@@ -114,7 +90,7 @@ public class RodNet : Module<Tensor, Tensor, Tensor, Tensor, Tensor>
         }
     }
 
-    internal static int GetRodState(RodInput input)
+    public virtual int GetRodState(RodInput input)
     {
         double[] pred = ComputeScores(input);
 
@@ -145,83 +121,6 @@ public class RodNet : Module<Tensor, Tensor, Tensor, Tensor, Tensor>
         pred[0] -= offset[fish_label]; // to make the prediction more precise when deployed
 
         return pred;
-    }
-
-    internal int GetRodState_Torch(RodInput input)
-    {
-        using var _ = no_grad();
-        Tensor outputTensor = ComputeScores_Torch(input);
-
-        var max = argmax(outputTensor);
-        return (int)max.item<long>();
-    }
-
-    public Tensor ComputeScores_Torch(RodInput input)
-    {
-        using var _ = no_grad();
-        this.SetWeightsManually();
-
-        var (y0, z0, t, u, v, h) = GetRodStatePreProcess(input);
-
-        Tensor fishLabel = tensor(new double[] { input.fish_label }, dtype: ScalarType.Int32);
-        Tensor uv = tensor(new double[,] { { u, v } }, dtype: ScalarType.Float64);
-        Tensor y0z0t = tensor(new double[,] { { y0, z0, t } }, dtype: ScalarType.Float64);
-        Tensor h_ = tensor(new double[,] { { h } }, dtype: ScalarType.Float64);
-
-        var logits = forward(fishLabel, uv, y0z0t, h_);
-        var output = PostProcess(logits, fishLabel);
-
-        return output;
-    }
-
-    /// <summary>
-    /// 使用时直接赋值已知权重
-    /// </summary>
-    public void SetWeightsManually()
-    {
-        var weightTensor = tensor(RodNet.weight, ScalarType.Float64);
-        var biasTensor = tensor(RodNet.bias, ScalarType.Float64);
-        var dzTensor = tensor(RodNet.dz, ScalarType.Float64).reshape([RodNet.dz.Length, 1]);
-        var h_coeffTensor = tensor(RodNet.h_coeff, ScalarType.Float64).reshape([RodNet.h_coeff.Length, 1]);
-        this.thetaParameter = new Parameter(weightTensor);
-        this.bParameter = new Parameter(biasTensor);
-        this.dzParameter = new Parameter(dzTensor);
-        this.hCoeffParameter = new Parameter(h_coeffTensor);
-    }
-
-    public override Tensor forward(Tensor fishLabel, Tensor uv, Tensor y0z0t, Tensor h)
-    {
-        var uvSplit = uv.split([1, 1], dim: 1);
-        Tensor u = uvSplit[0];
-        Tensor v = uvSplit[1];
-
-        var y0z0tSplit = y0z0t.split([1, 1, 1], dim: 1);
-        Tensor y0 = y0z0tSplit[0];
-        Tensor z0 = y0z0tSplit[1];
-        Tensor t = y0z0tSplit[2];
-
-        v = v - h * hCoeffParameter[fishLabel];
-
-        Tensor x, y, dist;
-
-        var dz = dzParameter[fishLabel];
-        x = u * (z0 + dz) * torch.sqrt(1 + t * t) / (t - v);
-        y = (z0 + dz) * (1 + t * v) / (t - v);
-        dist = torch.sqrt(x * x + (y - y0) * (y - y0));
-
-        Tensor logits = this.thetaParameter[fishLabel] * dist + this.bParameter[fishLabel];
-
-        return logits;
-    }
-
-    public Tensor PostProcess(Tensor logits, Tensor fishLabel)
-    {
-        var x_softmax = torch.nn.functional.softmax(logits, 1);
-
-        Tensor x_offset = tensor(fishLabel.data<int>().Select(l => RodNet.offset[l]).ToArray());
-
-        x_softmax[torch.arange(x_offset.shape[0]), 0] -= x_offset;
-        return x_softmax;
     }
 
     /// <summary>
