+++ BetterGenshinImpact/GameTask/Model/GameUI/ArtifactSetFilterScreen.cs
@@ -10,7 +10,6 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using TorchSharp.Modules;
 
 namespace BetterGenshinImpact.GameTask.Model.GameUI
 {
