+++ BetterGenshinImpact/User/I18n/en.json
@@ -234,7 +234,6 @@
   "钓全天的鱼、还是只钓白天或夜晚的鱼、亦或不调整时间": "Fish all day, only day or night fish, or do not adjust time",
   "关键帧保存截图（开发者）": "Save keyframe screenshots (developer)",
   "在流程判断的关键时刻保存当时的截图，供分析判断。会大量写入，非调试时请关闭。需要启用保存截图功能": "Save screenshots at key decision moments for analysis. Writes a lot—turn off when not debugging. Requires enabling screenshot saving",
-  "torch库文件地址（仅限2.5.1版本）": "Torch library file path (2.5.1 only)",
   "请 ": "Please ",
   "获取剪切板上的兑换码": "Get redemption code from clipboard",
   "在切换到软件界面时候，自动提取兑换码并提示": "When switching to the app window, automatically extract redemption codes and notify",
@@ -269,7 +268,6 @@
   "E战技": "Elemental Skill (E)",
   "Q爆发": "Elemental Burst (Q)",
   "检测盾奶位战技短按或长按，禁用:短按 / 启用:长按": "Detect shielder/healer slot skill short/long press. Disable: short press / Enable: long press",
-  " 到本地后填入torch_cpu.dll或torch_cuda.dll的完整地址。如未生效可尝试重启BGI。": "After downloading locally, enter the full path to torch_cpu.dll or torch_cuda.dll. If it doesn't take effect, try restarting BGI.",
   "下载": "Download",
   "一般填写套装内生之花名，可填入多个名称；留空则不用": "Usually enter the artifact set's Flower name; you can enter multiple names. Leave blank to disable",
   "自动检测战斗结束": "Auto detect combat end",
