+++ BetterGenshinImpact/User/I18n/ja.json
@@ -234,7 +234,6 @@
   "钓全天的鱼、还是只钓白天或夜晚的鱼、亦或不调整时间": "一日中釣りをするか、昼か夜だけ釣りをするか、調整しないでください時間",
   "关键帧保存截图（开发者）": "キーフレームのスクリーンショットを保存する (開発者)",
   "在流程判断的关键时刻保存当时的截图，供分析判断。会大量写入，非调试时请关闭。需要启用保存截图功能": "分析のために重要な決定の瞬間にスクリーンショットを保存します。大量の書き込みを行います。デバッグしないときはオフにします。スクリーンショットの保存を有効にする必要があります",
-  "torch库文件地址（仅限2.5.1版本）": "Torch ライブラリ ファイル パス (2.5.1 のみ)",
   "请 ": "",
   "获取剪切板上的兑换码": "クリップボードから引き換えコードを取得してください",
   "在切换到软件界面时候，自动提取兑换码并提示": "アプリ ウィンドウに切り替えると、自動的に引き換えコードを抽出して通知します",
@@ -269,7 +268,6 @@
   "E战技": "エレメンタル スキル (E)",
   "Q爆发": "エレメンタル バースト (Q)",
   "检测盾奶位战技短按或长按，禁用:短按 / 启用:长按": "シールドシールド/ヒーラー スロット スキルを短押し/長押しします。無効: 短押し / 有効: 長押し",
-  " 到本地后填入torch_cpu.dll或torch_cuda.dll的完整地址。如未生效可尝试重启BGI。": "ローカルにダウンロードした後、torch_cpu.dll または torch_cuda.dll へのフルパスを入力します。有効にならない場合は、BGI を再起動してみてください。",
   "下载": "ダウンロード",
   "一般填写套装内生之花名，可填入多个名称；留空则不用": "通常は、アーティファクト セットの花の名前を入力します。複数の名前を入力できます。無効にする場合は空白のままにします",
   "自动检测战斗结束": "戦闘終了の自動検出",
