+++ Test/BetterGenshinImpact.UnitTest/BetterGenshinImpact.UnitTest.csproj
@@ -16,6 +16,7 @@
         <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0"/>
         <PackageReference Include="xunit" Version="2.5.3"/>
         <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3"/>
+		<PackageReference Include="TorchSharp" Version="0.105.0" />
     </ItemGroup>
 
     <ItemGroup>
