+++ Test/BetterGenshinImpact.UnitTest/GameTaskTests/AutoFishingTests/TorchFixture.cs
@@ -1,5 +1,3 @@
-﻿using BetterGenshinImpact.GameTask.AutoFishing;
-using Microsoft.Extensions.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -26,13 +24,9 @@ namespace BetterGenshinImpact.UnitTest.GameTaskTests.AutoFishingTests
     {
         public TorchLoader()
         {
-            // 需要读取主项目编译目录中的配置
-            string configFullPath = Path.Combine(Path.GetFullPath(@"..\..\..\..\..\"), @"BetterGenshinImpact\bin\x64\Debug\net8.0-windows10.0.22621.0\User\config.json");
-            IConfigurationRoot configurationRoot = new ConfigurationBuilder().AddJsonFile(configFullPath, optional: false).Build();
-            AutoFishingConfig autoFishingConfig = configurationRoot.GetRequiredSection("autoFishingConfig").Get<AutoFishingConfig>() ?? throw new ArgumentNullException();
             try
             {
-                NativeLibrary.Load(autoFishingConfig.TorchDllFullPath);
+                NativeLibrary.Load(@"C:\torch\lib\torch_cpu.dll");
                 if (torch.TryInitializeDeviceType(DeviceType.CUDA))
                 {
                     torch.set_default_device(new torch.Device(DeviceType.CUDA));
